requirements.txt
pip~=24.2
build~=1.2.2
wheel~=0.44.0
docutils~=0.21.2
Jinja2~=3.1.3
keyring~=25.3.0
setuptools~=70.1.1
packaging~=23.2
zipp~=3.20.2
aiogram
python-dotenv
pywin32
pillow
google-api-python-client
google-auth-oauthlib
google-auth-httplib2